POG-236 [fix]: 큐를 등록하면, Task 데코레이터가 큐를 인식하여 큐에 작업 추가
